refactor connect button into another component
